Adds a rosbag_validate binary. (#536)
So far it creates statistics over timing information of the sensor data that Cartographer cares about. It also reports if time jumps backwards for a sensor.
Also updates the GitHub issue template to ask users to run this binary when reporting issues.

I verified that this tool is sufficient to analyze all the timing related issues mentioned in #529.