Extract assets writer class from static method (#827)
Extracted class Assets_Writer from RunAssetsWriterPipeline. 

The idea is to increase the re-usability and flexibility of the assets_writer: In next PR, the assets_writer will allow registering external points_processers to the points processing pipeline. This requires having a class instead of a static method to allow for different states.