Fix memory leak in simulations by removing WallTimers.
Fixes the problem of ever-growing memory after `rosbag play` finishes,
as discussed in https://github.com/googlecartographer/cartographer/issues/1182

The WallTimers caused the timer callback that publishes TF data to be called
even if no simulated `/clock` was published anymore.
As the TF buffer cache time of the TF listener seems to be based on
the ROS time instead of wall clock, it could grow out of bounds.

Now, `ros::Timer` plays nicely with both normal and simulated time and
no callbacks are executed if `/clock` stops in simulation.