Fix bug in FinishTrajectory logic
This PR adds additional bookkeeping for trajectories that we ischeduled for
finishing.

In Node::RunFinalOptimization(...), we were calling FinishTrajectory for
every trajectry with state == ACTIVE. Since the state only gets updated
once the corresponding worker for the FinishTrajectory task is
scheduled, we were potentially calling FinishTrajectory twice for a
single Trajectory id, if the worker has not yet completed and updated the state.