Add a script for testing with fake landmarks to scripts/dev
Can be used to test the landmark feature with fake, randomly sampled landmarks.

For example, it allows to tune the baseline publish period that
sometimes leads to confusion. We used this script for internal tests,
but it could be also useful for others (see #1067, #909).

For example:

```
./publish_fake_random_landmarks.py \
  --publish_period 0.1
  --id_vocabulary A B C
  --id_length 5
  --sample_period 1.0
```
will publish empty landmark lists at 10Hz and random landmarks every 1
seconds. IDs are also sampled, using the cartesian product of the provided
"vocabulary" - in the above example e.g. "AACBC" (length=5).