Remove unnecessary bottlenecks from Dockerfiles to speed up CI. (#1511)
This should solve #1506.

- Use osrf/ros-<distro>-desktop as base image for CI
The basic ROS base image requires rosdep to install packages with many
dependencies, e.g. rviz. Using the desktop base image should speed up CI
because it has a good overlap with what we need and only little extra stuff.
- Remove apt list only after installing everything.
- Don't build protobuf 3 from source on Ubuntu 18 & 20.
It should come with rosdep. Only on Ubuntu 16 we need to build manually
because it ships with Protobuf 2.
-  No need to build Ceres from source on Ubuntu 18 & 20. 
The versions are greater or equal to 1.13, which is the one specified in
the source build script. The apt packages should work just fine.

The cartographer library is built from source in all cases, which is good to
ensure that we always test compatibility with the latest version of it.

Signed-off-by: Guillaume Doisy <guillaume.doisy@wyca.fr>